requirements.txt
torch>=2.0.0
transformers>=4.30.0
scanpy>=1.9.0
datasets>=2.14.0
anndata>=0.9.0
scvi-tools>=1.0.0
mygene>=3.2.0
pandas>=1.5.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
joblib>=1.3.0
tqdm>=4.65.0
seaborn>=0.12.0 